nrpe_ng.commands: introduce tornado; use for process execution
The HTTP stuff will be replaced by Tornado soon, but a gentle
introduction of it in the process execution part seems like a good place
to start. This replaces some horrible code and pointless threading with
much more sensible event loop stuff.

This is not complete: killing processes on timeout won't work because we
only run the IOLoop for the one coroutine.