client: try to extract more friendly messages from requests exceptions
The real underlying exception that is generated when an HTTP request
fails is often wrapped in many layers of exceptions. For a simple
'Connection refused' error, the ConnectionRefusedError is wrapped in a
urllib3 NewConnectionError, which is itself wrapped in urllib3's
MaxRetryError, and finally a requests ConnectionError. This wouldn't be
so bad if there was a nice user-presentable string in the exception, or
some easy means of telling what happened: no, I get to peel back all the
layers of the onion myself.